TODOs por usuario

+++ src/todos/actions/todo-actions.ts
@@ -1,5 +1,6 @@
 'use server';
 
+import { getUserSessionServer } from "@/actions/auth-actions";
 import prisma from "@/lib/prisma";
 import { Todo } from "@prisma/client";
 import { revalidatePath } from "next/cache";
@@ -33,9 +34,9 @@ export const toggleTodo = async(id: string, complete: boolean): Promise<Todo> =>
     return updatedTodo;
 }
 
-export const addTodo = async(description: string) => {
+export const addTodo = async(description: string, userId: string) => {
     try {
-        const todo = await prisma.todo.create({data: {description}});
+        const todo = await prisma.todo.create({ data: {description, userId: userId}});
         revalidatePath('/dashboard/server-todos');
         return todo;
     } catch (error) {
@@ -46,6 +47,7 @@ export const addTodo = async(description: string) => {
 }
 
 export const deleteCompleted = async(): Promise<void> =>{
-    await prisma.todo.deleteMany({where: {complete: true}});
+    const user = await getUserSessionServer();
+    await prisma.todo.deleteMany({where: {complete: true, userId: user?.id}});
     revalidatePath('/dashboard/server-todos');
 }
